projects/example-project/requirements.txt
crewai>=0.1.0
crewai-tools>=0.1.0
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
